Rename StableSwapManage to StableFarmManage
