+++ src/pages/EarnTri/StableFarmManage.tsx
